Fix ninja rebuild loop and restore regressed functions
Fix case-sensitive path normalization in transform_dep.py that caused
SoundTouch files to rebuild every run (MSVC reports src\memory.h for
src/Memory.h).

Restore FreestyleMoveRecorder mRecordPos/mPlaybackPos to float (were
incorrectly retyped to int during rename). Fix EQEffect BiquadBand
struct layout back to flat fields (struct was wrong) and fix mBand5Freq
init value (25.0f not 0). Use local variable for CharServoBone assert.
Replace UISlider null-init loop with individual assignments.

+++ src/system/hamobj/FreestyleMoveRecorder.h
@@ -83,8 +83,8 @@ private:
     int mRecordingFrameCount;
     int mLastFrameIndex; // 0x20 - initialized to -1
     int mMaxFrames;
-    int mRecordPos;
-    int mPlaybackPos;
+    float mRecordPos;
+    float mPlaybackPos;
     float mDefaultTimeout; // 0x30 - initialized to 15
     int mPlaybackIndex; // 0x34 - initialized to -1, playback position
     bool mRecording;
